backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
python-multipart>=0.0.6
pygls>=1.1.0
neo4j>=5.12.0
python-jose>=3.3.0
python-dotenv>=1.0.0
git-python>=1.0.0
tree-sitter>=0.0.0
aiosqlite>=0.17.0
aiofiles>=0.7.0
sqlalchemy>=1.4.0
redis>=3.5.0
asyncpg>=0.29.0
uvicorn[standard]>=0.23.0